move search related logic to a context
